Move OrbitalObjectUtils to its own file
